Moved main application to use terminal component
