+++ frontend/src/index.jsx
@@ -2,8 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { BrowserRouter, Routes, Route } from "react-router";
 import App from "./App";
-import Test from "./Test";
-// import "./index.css";
+import "./index.css";
 
 const root = createRoot(document.getElementById("root"));
 root.render(
@@ -11,7 +10,6 @@ root.render(
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<App />} />
-        <Route path="/test" element={<Test />} />
       </Routes>
     </BrowserRouter>
   </StrictMode>,
